requirements.txt
PyYAML>=6.0
PyLaTeX>=1.4.1

Jinja2>=3.0
WeasyPrint>=50.0
